+++ R/plume.R
@@ -132,8 +132,11 @@ Plume <- R6Class(
     #' @return A character vector.
     get_author_list = function(suffix = NULL, format = deprecated()) {
       if (lifecycle::is_present(format)) {
-        lifecycle::deprecate_warn("0.2.0", "get_author_list(format)", "get_author_list(suffix)")
-        suffix <- format
+        lifecycle::deprecate_stop(
+          "0.2.1",
+          "get_author_list(format)",
+          "get_author_list(suffix)"
+        )
       }
       authors <- private$pull("literal_name")
       if (is_empty(suffix)) {