+++ NEWS.md
@@ -10,6 +10,8 @@
 
 * Defining explicit roles in the input data, deprecated in 0.2.0, is now defunct. Roles must now be specified using the `roles` parameter in `Plume$new()` and `PlumeQuarto$new()`.
 
+* The `format` parameter in `$get_author_list()`, deprecated in 0.2.1, is now defunct.
+
 # plume 0.2.5
 
 * `$get_author_list(suffix =)` is now slightly more flexible and will try less hard to sanitise author list suffixes (#90).